backend/requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.3
cryptography==42.0.5
deprecation==2.1.0
Django==5.0.4
django-celery-beat==2.6.0
django-dirtyfields==1.9.2
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
drf-extra-fields==3.7.0
fancycompleter==0.9.1
filetype==1.2.0
firebase-admin==6.5.0
flower==2.0.1
google-api-core==2.18.0
google-api-python-client==2.126.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
gotrue==2.4.2
grpcio==1.62.1
grpcio-status==1.62.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
humanize==4.9.0
idna==3.7
iniconfig==2.0.0
kombu==5.3.7
model-bakery==1.17.0
msgpack==1.0.8
packaging==24.0
pdbpp==0.10.3
pluggy==1.4.0
postgrest==0.16.3
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psycopg==3.1.18
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.2
pyrepl==0.9.0
pytest==8.1.1
pytest-django==4.8.0
pytest-env==1.1.3
python-crontab==3.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
realtime==1.0.4
redit==1.0.2
requests==2.31.0
responses==0.25.0
rsa==4.9
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
storage3==0.7.4
StrEnum==0.4.15
supabase==2.4.3
supafunc==0.4.5
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wait-for-it==2.2.2
wcwidth==0.2.13
websockets==12.0
whitenoise==6.6.0
wmctrl==0.5
